backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
boto3==1.35.91
celery==5.4.0
redis==5.2.1
sqlparse==0.5.5
tzdata==2025.3
